third_party/WebKit/LayoutTests/editing/deleting/smart_editing_disabled.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => {
  assert_own_property(window, 'internals', 'This test requires window.internals.');
  assert_own_property(window, 'testRunner', 'This test requires window.testRunner.');
  internals.settings.setSmartInsertDeleteEnabled(false);
  internals.settings.setSelectTrailingWhitespaceEnabled(true);
  assert_selection(
    '<div contenteditable>foo |bar^ baz</div>',
    seleciton => testRunner.execCommand('deleteForward'),
    '<div contenteditable>foo\u00A0| baz</div>',
    'DeleteForward without smart insert/delete');
}, 'Verifies deleteForward without smrat insert delete');
</script>
